+++ RELEASES.md
@@ -725,8 +725,7 @@
 -----------------
   - fix #274: Bring back IE11 compatibility.
 
-11.2.3 2020-03-05
------------------
+11.2.3 2020-03-05-----------------
   - fix: #272 - primus leaks on client __endSocket
   - fix: removed Promisify of handleMessage in session service
 
@@ -743,3 +742,7 @@
   - fix: 403 returned on token ttl, issue #287
   - feature: leaner publish functionality on the client, issue #281
   - refactor: CONSTANTS being used for authorization failure messages
+
+11.3.1 2020-04-20
+-----------------
+  - fix #292 - eat the rest of the post after sending an unauthorized/forbidden